refactor(idp-better-auth): refactor types
